Codigo terminado pero sin diseños bonitos

+++ proyecto-web/app/Http/Controllers/CarritoController.php
@@ -60,26 +60,30 @@ class CarritoController extends Controller
 
     public function confirmarCarrito(){
         //dd(Cart::subtotal());
-        $pedido = new Pedido();
-        $pedido->subtotal = floatval(str_replace(',', '', Cart::subtotal()));
-        $pedido->impuesto = floatval(str_replace(',', '', Cart::tax()));
-        $pedido->total = floatval(str_replace(',', '', Cart::total()));
-        $pedido->fechapedido = Carbon::now();
-        $pedido->estado = "nuevo"; 
-        $pedido->user_id= auth()->user()->id;
-        $pedido->save();
+        if(Cart::count() > 0){
+            $pedido = new Pedido();
+            $pedido->subtotal = floatval(str_replace(',', '', Cart::subtotal()));
+            $pedido->impuesto = floatval(str_replace(',', '', Cart::tax()));
+            $pedido->total = floatval(str_replace(',', '', Cart::total()));
+            $pedido->fechapedido = Carbon::now();
+            $pedido->estado = "nuevo"; 
+            $pedido->user_id= auth()->user()->id;
+            $pedido->save();
 
-        foreach(Cart::content() as $item){
-            $detalle = new Detalle();
-            $detalle->precio = $item->price;
-            $detalle->importe = $item->price * $item->qty;
-            $detalle->cantidad = $item->qty;
-            $detalle->pedido_id = $pedido->id;
-            $detalle->producto_id = $item->id;
-            $detalle->save();
+            foreach(Cart::content() as $item){
+                $detalle = new Detalle();
+                $detalle->precio = $item->price;
+                $detalle->importe = $item->price * $item->qty;
+                $detalle->cantidad = $item->qty;
+                $detalle->pedido_id = $pedido->id;
+                $detalle->producto_id = $item->id;
+                $detalle->save();
+            }
+
+            Cart::destroy();
+            return redirect()->route('index')->with('success','Pedido realizado correctamente');
+        }else{
+            return redirect()->route('index')->with('error','No hay productos en el carrito');
         }
-
-        Cart::destroy();
-        return redirect()->route('index')->with('success','Pedido realizado correctamente');
     }
 }
